Rename ParameterMetaDataIO -> ParameterMetaDataSyntax
